convocatoria crear

+++ database/migrations/2025_04_04_230335_convocatoria_area_categoria.php
@@ -13,7 +13,7 @@ class ConvocatoriaAreaCategoria extends Migration
             $table->unsignedBigInteger('idConvocatoria');
             $table->unsignedBigInteger('idArea');
             $table->unsignedBigInteger('idCategoria');
-            $table->integer('precio');
+            $table->decimal('precio', 8, 2); // Changed from integer to decimal
             $table->timestamps();
 
             $table->foreign('idConvocatoria')->references('idConvocatoria')->on('convocatoria')->onDelete('cascade');
